pyproject.toml
[project]
name = "kann-ai-bot"
version = "0.1.0"
description = "WikiHow-basierte KI-Fragen fuer Mastodon und Web"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "httpx>=0.28",
    "mastodon.py>=1.8",
    "python-dotenv>=1.0",
    "aiosqlite>=0.20",
    "python-multipart>=0.0.20",
    "pillow>=12.1.1",
    "slowapi>=0.1.9",
    "atproto>=0.0.65",
]

[tool.hatch.build.targets.wheel]
packages = ["web"]
